LF-5213 Resolve circular reference
